+++ docs/changelog.rst
@@ -2,6 +2,8 @@
 Changelog
 =========
 
+- :feature:`-` Added the ``invocations.environment`` module with top-level
+  functions such as `~invocations.environment.in_ci`.
 - :release:`2.1.0 <2021-08-27>`
 - :feature:`-` Add ``packaging.release.test_install`` task and call it just
   prior to the final step in ``packaging.release.upload`` (so one doesn't